databaser fixet, tabeller oprettet.PROGRAMMET STARTER 14.14

+++ src/main/java/com/example/kinoxpbackend/repository/CinemaRepository.java
@@ -8,5 +8,5 @@ import java.awt.*;
 
 
 @Repository
-public interface CinemaRepository extends JpaRepository<Cinema,Long> {
+public interface CinemaRepository extends JpaRepository<Cinema,Integer> {
 }
